% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_CB.R
\name{pi_CB}
\alias{pi_CB}
\title{Prediction intervals based on the CB algorithm}
\usage{
pi_CB(series, p = 1, h = 3, B = 1000, alpha = 0.05, method = c("OLS", "LAD"))
}
\arguments{
\item{series}{Numeric vector or ts object with time series values.}

\item{p}{Positive integer indicating the autoregressive order. Default is 1.}

\item{h}{Positive integer indicating the prediction horizon. Default is 3.}

\item{B}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Significance level. 1-\code{alpha} is the nominal coverage level. Default is 0.05.}

\item{method}{Estimation method. One among "OLS" and "LAD". Default is "OLS".}
}
\value{
A list with elements:
\describe{
\item{bfor}{Numeric matrix of bootstrap forecasts with dimension h x B.}
\item{lpi}{Numeric vector of lower bounds (length \code{h}).}
\item{upi}{Numeric vector of upper bounds (length \code{h}).}
\item{len}{Numeric vector of interval lengths (length \code{h}).}
}
}
\description{
Computes bootstrap percentile-based prediction intervals using the
conditional bootstrap algorithm of Cao et al. (1997) for
an AR(p) model over a forecast horizon \code{h}.
}
\details{
This function implements the conditional bootstrap algorithm described in
Cao et al. (1997).
}
\examples{
set.seed(123)

# Simulation parameters
burn_in <- 300
n <- 25
m <- burn_in + n
coeff <- 0.6

# Simulate AR(1) process
e <- rnorm(m)
x <- numeric(m)
x[1] <- rnorm(1)

for (t in 2:m) {
  x[t] <- coeff * x[t - 1] + e[t]
}

series <- ts(x[(burn_in + 1):m])

# Compute prediction interval
out <- pi_CB(series,h=4,alpha=0.01)

out$lpi
out$upi
out$len

}
\references{
Cao, R., Febrero-Bande, M., González-Manteiga, W., Prada-Sánchez, J., and García-Jurado, I. (1997).
Saving computer time in constructing consistent bootstrap prediction intervals for autoregressive
processes. \emph{Communications in Statistics-Simulation and Computation}, 26(3):961–978.
}
