% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_PP.R
\name{pi_PP}
\alias{pi_PP}
\title{Prediction intervals based on PP algorithm}
\usage{
pi_PP(series, p = 1, h = 3, B = 1000, alpha = 0.05)
}
\arguments{
\item{series}{Numeric vector or ts object with time series values.}

\item{p}{Positive integer indicating the autoregressive order. Default is 1.}

\item{h}{Positive integer indicating the prediction horizon. Default is 3.}

\item{B}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Significance level. 1-\code{alpha} is the nominal coverage level. Default is 0.05.}
}
\value{
A list with elements:
\describe{
\item{pfor}{Numeric vector of point forecasts (length h).}
\item{lpi}{Numeric vector of lower bounds (length \code{h}).}
\item{upi}{Numeric vector of upper bounds (length \code{h}).}
\item{len}{Numeric vector of interval lengths (length \code{h}).}
}
}
\description{
Computes a bootstrap predictive-root-based prediction interval using the forward algorithm
of Pan and Politis (2016) for an AR(p) model over a forecast horizon \code{h}.
}
\details{
This function implements the PP algorithm of Pan and Politis (2016),
referred to in their article as Fp (forward bootstrap with predictive residuals).
}
\examples{
set.seed(123)

# Simulation parameters
burn_in <- 300
n <- 25
m <- burn_in + n
coeff <- 0.6

# Simulate AR(1) process
e <- rnorm(m)
x <- numeric(m)
x[1] <- rnorm(1)

for (t in 2:m) {
  x[t] <- coeff * x[t - 1] + e[t]
}

series <- ts(x[(burn_in + 1):m])

# Compute prediction interval
out <- pi_PP(series,alpha=0.10)

out$lpi
out$upi
out$len

}
\references{
Pan, L. and Politis, D. N. (2016). Bootstrap prediction intervals for linear, nonlinear and nonpara-
metric autoregressions. \emph{Journal of Statistical Planning and Inference}, 177:1–27.
}
