% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_PRR.R
\name{pi_PRR}
\alias{pi_PRR}
\title{Prediction intervals based on the PRR algorithm}
\usage{
pi_PRR(series, p = 1, h = 3, B = 1000, alpha = 0.05, method = c("LAD", "OLS"))
}
\arguments{
\item{series}{Numeric vector or ts object with time series values.}

\item{p}{Positive integer indicating the autoregressive order. Default is 1.}

\item{h}{Positive integer indicating the prediction horizon. Default is 3.}

\item{B}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Significance level. 1-\code{alpha} is the nominal coverage level. Default is 0.05.}

\item{method}{Estimation method. One among "OLS" and "LAD". Default is "LAD".}
}
\value{
A list with elements:
\describe{
\item{bfor}{Numeric matrix of bootstrap forecasts with dimension h x B.}
\item{lpi}{Numeric vector of lower bounds (length \code{h}).}
\item{upi}{Numeric vector of upper bounds (length \code{h}).}
\item{len}{Numeric vector of interval lengths (length \code{h}).}
}
}
\description{
Computes bootstrap percentile-based prediction intervals using the forward algorithm of
Pascual, Romo and Ruiz (2004) for an AR(p) model over a
forecast horizon \code{h}.
}
\details{
This function implements the forward bootstrap algorithm described in
Pascual, Romo and Ruiz (2004).
}
\examples{
set.seed(123)

# Simulation parameters
burn_in <- 300
n <- 25
m <- burn_in + n
coeff <- 0.6

# Simulate AR(1) process
e <- rnorm(m)
x <- numeric(m)
x[1] <- rnorm(1)

for (t in 2:m) {
  x[t] <- coeff * x[t - 1] + e[t]
}

series <- ts(x[(burn_in + 1):m])

# Compute prediction interval
out <- pi_PRR(series,h=4)

out$lpi
out$upi
out$len

}
\references{
Pascual, L., Romo, J., and Ruiz, E. (2004). Bootstrap predictive inference for ARIMA processes.
\emph{Journal of Time Series Analysis}, 25(4):449–465.
}
