% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_QAR_proot.R
\name{pi_QAR_proot}
\alias{pi_QAR_proot}
\title{Prediction intervals based on QAR-proot algorithm}
\usage{
pi_QAR_proot(series, p = 1, h = 3, B = 1000, alpha = 0.05, tau = 0.5)
}
\arguments{
\item{series}{Numeric vector or ts object with time series values.}

\item{p}{Positive integer indicating the autoregressive order. Default is 1.}

\item{h}{Positive integer indicating the prediction horizon. Default is 3.}

\item{B}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Significance level. 1-\code{alpha} is the nominal coverage level. Default is 0.05.}

\item{tau}{Quantile level used in estimation. Default is 0.5.}
}
\value{
A list with elements:
\describe{
\item{pfor}{Numeric vector of point forecasts (length h).}
\item{lpi}{Numeric vector of lower bounds (length \code{h}).}
\item{upi}{Numeric vector of upper bounds (length \code{h}).}
\item{len}{Numeric vector of interval lengths (length \code{h}).}
}
}
\description{
Computes a bootstrap predictive-root-based prediction interval using the QAR-proot algorithm for
a QAR(p) model over a forecast horizon \code{h}.
}
\details{
This function implements the QAR-proot algorithm described in
Novo and Sánchez-Sellero (2025).
}
\examples{
set.seed(123)

# Simulation parameters
burn_in <- 300
n <- 25
m <- burn_in + n
coeff <- 0.6

# Simulate AR(1) process
e <- rnorm(m)
x <- numeric(m)
x[1] <- rnorm(1)

for (t in 2:m) {
  x[t] <- coeff * x[t - 1] + e[t]
}

series <- ts(x[(burn_in + 1):m])

# Compute prediction interval
out <- pi_QAR_proot(series)

out$lpi
out$upi
out$len

#' # Simulate QAR(1) process
n <- 100
e<-runif(m)
x <- numeric(m)
x[c(1,2)] <- rt(2,3)

for (t in 3:m) {x[t]<-0.3*x[t-1]+0.7*e[t]*x[t-2]+qt(e[t],3)}
series2<-ts(x[301:m])

# Compute prediction interval

out2<-pi_QAR_proot(series2,h=1)
out2$lpi
out2$upi
out2$len

}
\references{
Novo, S., & Sánchez-Sellero, C. (2025).
\emph{Prediction intervals for quantile autoregression}.
arXiv:2512.22018. \url{https://arxiv.org/abs/2512.22018}
}
