% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_TS.R
\name{pi_TS}
\alias{pi_TS}
\title{Prediction intervals based on the TS algorithm}
\usage{
pi_TS(series, p = 1, h = 3, B = 1000, alpha = 0.05)
}
\arguments{
\item{series}{Numeric vector or ts object with time series values.}

\item{p}{Positive integer indicating the autoregressive order. Default is 1.}

\item{h}{Positive integer indicating the prediction horizon. Default is 3.}

\item{B}{Number of bootstrap replicates. Default is 1000.}

\item{alpha}{Significance level. 1-\code{alpha} is the nominal coverage level. Default is 0.05.}
}
\value{
A list with elements:
\describe{
\item{bfor}{Numeric matrix of bootstrap forecasts with dimension h x B.}
\item{lpi}{Numeric vector of lower bounds (length \code{h}).}
\item{upi}{Numeric vector of upper bounds (length \code{h}).}
\item{len}{Numeric vector of interval lengths (length \code{h}).}
}
}
\description{
Computes bootstrap percentile-based prediction intervals using the backward
algorithm of Thombs and Schucany (1990) for an AR(p) model over a
forecast horizon \code{h}.
}
\details{
This function implements the backward bootstrap algorithm proposed by
Thombs and Schucany (1990).
}
\examples{
set.seed(123)

# Simulation parameters
burn_in <- 300
n <- 25
m <- burn_in + n
coeff <- 0.6

# Simulate AR(1) process
e <- rnorm(m)
x <- numeric(m)
x[1] <- rnorm(1)

for (t in 2:m) {
  x[t] <- coeff * x[t - 1] + e[t]
}

series <- ts(x[(burn_in + 1):m])

# Compute prediction interval
out <- pi_TS(series,h=4)

out$lpi
out$upi
out$len

}
\references{
Thombs, L. A. and Schucany, W. R. (1990). Bootstrap prediction intervals for autoregression.
\emph{Journal of the American Statistical Association}, 85(410):486–492.
}
